connected to backend, done adding event, added Calendar and events table to database
