+++ server/calendar/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+bp = Blueprint("calendar", __name__)
+from . import routes
