+++ src/ components/authentication/AuthContext.jsx
@@ -23,7 +23,7 @@ export const AuthContextProvider = ({ children }) => {
     };
 
     const getUserProfile = async () => {
-        let apiResponse = await axios.get("http://127.0.0.1:5000/dashboard", {
+        let apiResponse = await axios.get("http://127.0.0.1:5000//api/user-profile", {
             withCredentials: true,
         });
         localStorage.setItem("userProfile", JSON.stringify(apiResponse.data.message));
@@ -32,24 +32,26 @@ export const AuthContextProvider = ({ children }) => {
     }
 
     const register = async (payload) => {
-        await axios.post("http://127.0.0.1:5000/users/auth/register", payload, {
-            withCredentials: true,
-        });
-        await getUserProfile();
-    }
+        try {
+            const response = await axios.post("http://127.0.0.1:5000/users/auth/register", payload, {
+                withCredentials: true,
+            });
+            await getUserProfile();
+            
+            return response.data.message; // Return the response data
+        } catch (error) {
+            return error;
+        }
+    };
 
     const login = async (payload) => {
         try {
             const response = await axios.post("http://127.0.0.1:5000/users/auth/login", payload, {
-                headers: {
-                    "Content-Type": "application/json",
-                },
                 withCredentials: true,
-                credentials: 'include',
             });
             console.log(response.data.message)
             await getUserProfile();
-            return response;
+            return response.data.message;
         }
         catch (error) {
             return error
