+++ src/App.jsx
@@ -1,27 +1,38 @@
 import  Timer  from './ components/timer/Timer';
 import HomePage from './ components/authentication/HomePage'
-import Calendar from './ components/calendar/Calendar';
-import { Route, Routes } from "react-router-dom";
+import { Navigate, Route, Routes } from "react-router-dom";
 import {AuthContextProvider} from "./ components/authentication/AuthContext.jsx";
 import ProtectedRoute from "./routes/ProtectedRoute.jsx"
+import SecondPage from './ components/second_page/SecondPage';
 
 export default function App() {
+  let isAuthenticated = localStorage.getItem("userProfile");
   return (
     <>
         <AuthContextProvider>
           <Routes>
-            <Route path="/" element={
-                  <ProtectedRoute accessBy="non-authenticated">
-                    <HomePage/>
-                  </ProtectedRoute>
-                  }>
-            </Route>
-            <Route path="/dashboard" element={
-                  <ProtectedRoute accessBy="authenticated">
-                    <DashboardLayout/>
-                  </ProtectedRoute>
-                  }>
-            </Route>
+            <Route
+              path="/"
+              element={
+                isAuthenticated ? (
+                  <Navigate to="/dashboard" />
+                ) : (
+                  <HomePage />
+                )
+              }
+            />
+            <Route
+            path="/dashboard"
+            element={
+              isAuthenticated ? (
+                <ProtectedRoute accessBy="authenticated">
+                  <DashboardLayout />
+                </ProtectedRoute>
+              ) : (
+                <Navigate to="/" />
+              )
+            }
+          />
           </Routes>
         </AuthContextProvider>
     </>
@@ -31,7 +42,7 @@ function DashboardLayout() {
   return (
     <div>
       <Timer />
-      <Calendar />
+      <SecondPage />
     </div>
     );
 }
