+++ src/assets/styles/calendar.css
@@ -1,14 +1,13 @@
 .calendar--wrapper {
-    min-height: 100vh;
     display: flex;
     align-items: center;
     justify-content: center;
+    margin: 1.5rem auto;
 }
 .calendar--container {
     position: relative;
     width: 1200px;
-    min-height: 850px;
-    margin: 0 auto;
+    /* margin: 5rem auto; */
     padding: 5px;
     color: #fff;
     display: flex;
