+++ src/assets/styles/secondPage.css
@@ -1,0 +1,6 @@
+.second-page__container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;    
+}
